requirements.txt
agora-token-builder
asgiref
attrs
autobahn
Automat
certifi
cffi
channels
channels_redis
charset-normalizer
constantly
cryptography
daphne
defusedxml
hyperlink
idna
incremental
mnemonic
psycopg2-binary
pyasn1
pyasn1-modules
pycparser
PyJWT
pyOpenSSL
python-dotenv
python-magic
python3-openid
pytz
requests
requests-oauthlib
service-identity
six
sqlparse
Twisted
txaio

tzdata
urllib3
uvicorn
zope.interface

# Server
gunicorn

# Django
Django
django-stubs[compatible-mypy]
django-allauth
django-bootstrap-datepicker-plus
django-ckeditor
django-cleanup
django-crispy-forms
django-debug-toolbar
django-imagekit
django-js-asset
django-protected-media
django-tables2
django-thumbnails

# Image Manipulation
Pillow

# AAA
oauthlib

# Build System Compoents
types-setuptools

# Formatting/linting
flake8
black
mypy
isort

# Logging
structlog

# Useful debugging tools
ipython

# Data Modeling
pydantic
pydantic-core

# Type hinting
types-python-dateutil
typing-extensions